done with email & passeword validation and database pesistance
