--- templateSpring/templateSpring/src/main/java/com/Template/templateSpring/Service/Service.java
@@ -1,4 +1,0 @@
-package com.Template.templateSpring.Service;
-
-public interface Service {
-}
