+++ templateSpring/templateSpring/src/main/java/com/Template/templateSpring/TemplateSpringApplication.java
@@ -5,7 +5,6 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
 public class TemplateSpringApplication {
-
 	public static void main(String[] args) {
 		SpringApplication.run(TemplateSpringApplication.class, args);
 	}
