+++ templateSpring/templateSpring/src/main/java/com/Template/templateSpring/dto/UserSignUpDto.java
@@ -1,4 +1,4 @@
-package com.Template.templateSpring.Dto;
+package com.Template.templateSpring.dto;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotEmpty;
@@ -8,18 +8,9 @@ public class UserSignUpDto {
     @NotEmpty
     @NotBlank
     private String email;
-
     @NotEmpty
     @NotBlank
     private String password;
-
-    public String getConfirmationPassword() {
-        return confirmationPassword;
-    }
-
-    public void setConfirmationPassword(String confirmationPassword) {
-        this.confirmationPassword = confirmationPassword;
-    }
 
     private String confirmationPassword;
 
@@ -42,4 +33,9 @@ public class UserSignUpDto {
     public void setPassword(String password) {
         this.password = password;
     }
+
+    public String getConfirmationPassword() {
+        return confirmationPassword;
+    }
+
 }
