+++ templateSpring/templateSpring/src/main/java/com/Template/templateSpring/repository/UserRepository.java
@@ -1,11 +1,9 @@
-package com.Template.templateSpring.Repository;
+package com.Template.templateSpring.repository;
 
-import com.Template.templateSpring.Entity.User;
+import com.Template.templateSpring.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.util.Optional;
-
 public interface UserRepository extends JpaRepository<User, Integer> {
-    Optional<User> findByEmail(String email);
+    User findByEmail(String email);
     Boolean existsByEmail(String email);
 }
