+++ templateSpring/templateSpring/src/main/java/com/Template/templateSpring/service/Service.java
@@ -1,0 +1,13 @@
+//package com.Template.templateSpring.Service;
+//
+//import com.Template.templateSpring.Dto.UserSignUpDto;
+//import com.Template.templateSpring.Entity.User;
+//
+//import java.util.List;
+//
+//public interface Service {
+//    void saveUser(UserSignUpDto userSignUpDto);
+//    User findUserByEmail(String userEmail);
+//
+//    List<UserSignUpDto> findAllUsers();
+//}
