+++ templateSpring/templateSpring/src/main/java/com/Template/templateSpring/validator/EmailValidator.java
@@ -1,6 +1,6 @@
-package com.Template.templateSpring.Validator;
+package com.Template.templateSpring.validator;
 
-import com.Template.templateSpring.Dto.UserSignUpDto;
+import com.Template.templateSpring.dto.UserSignUpDto;
 
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
@@ -8,7 +8,7 @@ import java.util.regex.Pattern;
 public class EmailValidator{
     private Pattern pattern;
     private Matcher matcher;
-    private static final String EMAIL_PATTERN = "^(?=.{1,50}$)[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\\\\.[a-zA-Z]{2,}$";
+    private static final String EMAIL_PATTERN = "^(?=.{1,50}$)[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\\.[a-zA-Z]{2,}$";
     public boolean validateEmail(UserSignUpDto user) {
         pattern = Pattern.compile(EMAIL_PATTERN);
         matcher = pattern.matcher(user.getEmail());
