+++ templateSpring/templateSpring/src/main/resources/application.properties
@@ -1,0 +1,7 @@
+spring.datasource.url=jdbc:postgresql://localhost:5432/template1
+spring.datasource.username=postgres
+spring.datasource.password=mzwili@postgres1
+spring.jpa.hibernate.ddl-auto=update
+spring.jpa.show-sql=true
+spring.jpa.properties.hibernate.format_sql=true
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
