+++ templateSpring/templateSpring/src/main/resources/application.yml
@@ -1,8 +1,8 @@
-auth0:
-  audience: http://localhost:8080/signup/
-spring:
-  security:
-    oauth2:
-      resourceserver:
-        jwt:
-          issuer-uri: https://dev-a2y3pt18mov2zkwu.us.auth0.com/
+#auth0:
+#  audience: http://localhost:8080/signup/
+#spring:
+#  security:
+#    oauth2:
+#      resourceserver:
+#        jwt:
+#          issuer-uri: https://dev-a2y3pt18mov2zkwu.us.auth0.com/
